+++ sky-take-out/sky-server/src/main/java/com/sky/service/DishService.java
@@ -53,5 +53,5 @@ public interface DishService {
      * @param categoryId
      * @return
      */
-    List<Dish> queryByCategory(Integer categoryId);
+    List<Dish> queryByCategory(Long categoryId);
 }
